Improve CheckedMessage ergonomics
