+++ logger_test.go
@@ -30,7 +30,6 @@ import (
 	"github.com/uber-go/zap/spywrite"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func opts(opts ...Option) []Option {
@@ -47,7 +46,10 @@ func withJSONLogger(t testing.TB, opts []Option, f func(*jsonLogger, func() []st
 	jl := NewJSON(allOpts...)
 	jl.StubTime()
 
-	f(jl.(*jsonLogger), func() []string { return strings.Split(sink.String(), "\n") })
+	f(jl.(*jsonLogger), func() []string {
+		output := strings.Split(sink.String(), "\n")
+		return output[:len(output)-1]
+	})
 	assert.Empty(t, errSink.String(), "Expected error sink to be empty")
 }
 
@@ -192,7 +194,7 @@ func TestJSONLoggerNoOpsDisabledLevels(t *testing.T) {
 	withJSONLogger(t, nil, func(jl *jsonLogger, output func() []string) {
 		jl.SetLevel(Warn)
 		jl.Info("silence!")
-		assert.Equal(t, []string{""}, output(), "Expected logging at a disabled level to produce no output.")
+		assert.Equal(t, []string{}, output(), "Expected logging at a disabled level to produce no output.")
 	})
 }
 
@@ -249,28 +251,4 @@ func TestJSONLoggerSyncsOutput(t *testing.T) {
 
 	assert.Panics(t, func() { logger.Panic("foo") }, "Expected panic when logging at Panic level.")
 	assert.True(t, sink.Called(), "Expected logging at panic level to Sync underlying WriteSyncer.")
-}
-
-func TestJSONLoggerCheck(t *testing.T) {
-	withJSONLogger(t, opts(Info), func(jl *jsonLogger, output func() []string) {
-		assert.Nil(t, jl.Check(Debug, "Debug."), "Expected a nil CheckedMessage at disabled levels.")
-
-		cm := jl.Check(Info, "Info.")
-		require.NotNil(t, cm, "Expected non-nil CheckedMessage at enabled levels.")
-		cm.Write(Int("magic", 42))
-		assert.Equal(
-			t,
-			`{"msg":"Info.","level":"info","ts":0,"fields":{"magic":42}}`,
-			output()[0],
-			"Unexpected output after writing a CheckedMessage.",
-		)
-	})
-}
-
-func TestCheckedMessageIsSingleUse(t *testing.T) {
-	withJSONLogger(t, nil, func(jl *jsonLogger, output func() []string) {
-		cm := jl.Check(Info, "Single-use.")
-		cm.Write()
-		assert.Panics(t, func() { cm.Write() }, "Expected re-using a CheckedMessage to panic.")
-	})
 }
